requirements.txt
requests
pyyaml
pandas
numpy
jinja2
click
python-dateutil
structlog
beautifulsoup4
lxml
python-dotenv
yfinance
scikit-learn
lightgbm
pyarrow
